+++ examples/03_noinit/src/App.tsx
@@ -2,7 +2,11 @@ import React, { Suspense } from 'react';
 
 import { ErrorBoundary } from 'react-hooks-fetch';
 
+import { store } from './fetchStore';
 import Item from './Item';
+
+store.prefetch('1');
+store.prefetch('2');
 
 const renderError = (error: Error) => (
   <div>
@@ -14,11 +18,13 @@ const renderError = (error: Error) => (
 const App: React.FC = () => (
   <ErrorBoundary fallback={renderError}>
     <Suspense fallback={<span>Loading...</span>}>
+      <Item initialId="1" />
+      <hr />
+      <Item initialId="2" />
+      <hr />
       <Item />
       <hr />
-      <Item initialId="8" />
-      <hr />
-      <Item initialId="9" />
+      <Item />
     </Suspense>
   </ErrorBoundary>
 );