+++ examples/03_noprefetch/src/fetchFuncs.ts
@@ -1,3 +1,0 @@
-export type Result = { data: { first_name: string } };
-
-export const fetchFunc = async (userId: string) => (await fetch(`https://reqres.in/api/users/${userId}?delay=3`)).json();